% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{graph_policy_comparisons}
\alias{graph_policy_comparisons}
\title{Outputting line graphs comparing policy scenarios over time}
\usage{
graph_policy_comparisons(
  .data,
  baseline_var,
  comparison_var,
  group_var = grp_all,
  y_lab = ""
)
}
\arguments{
\item{baseline_var}{Column to graph baseline}

\item{comparison_var}{Column to graph comparison}

\item{group_var}{Column with \code{TRUE} values corresponding to group}

\item{y_lab}{Title for y axis}
}
\value{
A ggplot object
}
\description{
Outputting line graphs comparing policy scenarios over time
}
\examples{
out_data |>
  graph_policy_comparisons(out_ghq_baseline, out_ghq_reform,  y_lab = "GQH score")
}
